feat(setup): post site-info when setup site
