+++ pages/index/index.vue
@@ -3,6 +3,7 @@
 
 <template>
   <div class="mx-auto flex min-h-full w-full flex-col pb-2">
+    <nuxt-loading-indicator />
     <client-only>
       <navbar />
     </client-only>
